Resolve "Model versioning"

+++ apps/api/sentiment/model.py
@@ -14,6 +14,7 @@ import os
 PATH_TO_MODEL = Path("model")
 PATH_TO_OLD_MODEL = Path("model_old")
 
+
 # Pre-processes the text so that it can be used in the Machine Learning pipeline.
 # This is done by lowercasing the text, removing punctuation and numbers, stopwords, and finally lemmatizing the text.
 def preprocess_text(text):
@@ -28,17 +29,18 @@ def preprocess_text(text):
     text = " ".join(lemmatizer.lemmatize(word) for word in text.split())
     return text
 
+
 def train_model():
     db_conn = get_db_connection()
     train_df = pd.read_sql_query("SELECT * FROM api_sentimententry", db_conn)
-    
+
     print("Preprocessing data set")
     train_df["ProcessedTweet"] = train_df["text"].apply(preprocess_text)
 
     print("Vectorizing data set")
     vectorizer = CountVectorizer()
-    X = vectorizer.fit_transform(train_df['ProcessedTweet'])
-    y = train_df['sentiment']
+    X = vectorizer.fit_transform(train_df["ProcessedTweet"])
+    y = train_df["sentiment"]
 
     print("Training final model on the entire dataset")
     # Logistic Regression with K-Fold Cross-Validation
@@ -50,7 +52,7 @@ def train_model():
 
     # Initialize LR model
     # Tune hyperparameters (C)
-    final_model = LogisticRegression(max_iter=2000, C=3.0)  
+    final_model = LogisticRegression(max_iter=2000, C=3.0)
 
     # Train the model
     final_model.fit(X_train, y_train)
@@ -65,8 +67,11 @@ def train_model():
     if os.path.isfile(PATH_TO_OLD_MODEL):
         os.remove(PATH_TO_OLD_MODEL)
 
-    os.rename(PATH_TO_MODEL, PATH_TO_OLD_MODEL)
+    if os.path.isfile(PATH_TO_MODEL):
+        os.rename(PATH_TO_MODEL, PATH_TO_OLD_MODEL)
+
     joblib.dump(models, PATH_TO_MODEL)
+
 
 def preprocess_and_predict(text):
     # Load the model
@@ -87,12 +92,13 @@ def preprocess_and_predict(text):
 
     return lr_prediction
 
-def test_model(use_old_model = False):
+
+def test_model(use_old_model=False):
     db_conn = get_db_connection()
     test_df = pd.read_sql_query("SELECT * FROM api_sentimententrytest", db_conn)
     test_df = test_df.dropna()
-    test_df = test_df[['sentiment', 'text']]
-    test_df['ProcessedTweet'] = test_df['text'].apply(preprocess_text)
+    test_df = test_df[["sentiment", "text"]]
+    test_df["ProcessedTweet"] = test_df["text"].apply(preprocess_text)
 
     # Load the model and vectorizer
     try:
@@ -104,8 +110,8 @@ def test_model(use_old_model = False):
     model = models["model"]
 
     # Vectorize the test data
-    X_test = vectorizer.transform(test_df['ProcessedTweet'])
-    y_test = test_df['sentiment']
+    X_test = vectorizer.transform(test_df["ProcessedTweet"])
+    y_test = test_df["sentiment"]
 
     # Make predictions
     predictions_test = model.predict(X_test)
@@ -114,8 +120,9 @@ def test_model(use_old_model = False):
     accuracy = accuracy_score(y_test, predictions_test)
     return accuracy
 
+
 def predict_proba(texts):
-     # Load the model and vectorizer
+    # Load the model and vectorizer
     try:
         models = joblib.load(PATH_TO_MODEL)
     except FileNotFoundError:
@@ -128,5 +135,5 @@ def predict_proba(texts):
 
     # Predict probabilities for each feature.
     probabilities = model.predict_proba(vectorized_texts)
-    
+
     return probabilities
